+++ src/main/java/io/github/paritoshgpt1/Housie/repository/TicketRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface TicketRepository extends CrudRepository<Ticket, Long> {
+public interface TicketRepository extends CrudRepository<Ticket, Integer> {
     List<Ticket> findAllByPlayer(Player player);
 }
